[rl] Add unit test for fused parameter in-place write bug
Add TestFusedQKVInplaceWrite to the existing test_fused_qkv.py suite.
Demonstrates that FusedQKVLinear's state_dict post-hook returns
.contiguous() copies that don't share storage with the fused wqkv
parameter. In-place writes (as TorchStore's get_state_dict does) into
these copies leave the model unchanged unless load_state_dict is called
afterward to trigger the merge hooks.
